Finished refactor.
