+++ tests/providers/test_context_resources.py
@@ -11,17 +11,16 @@ from unittest.mock import Mock
 import pytest
 
 from that_depends import BaseContainer, Provide, fetch_context_item, inject, providers
-from that_depends.entities.context import _enter_named_scope
-from that_depends.entities.resource_context import ResourceContext
-from that_depends.meta import DefaultScopeNotDefinedError
-from that_depends.providers import container_context
-from that_depends.providers.context_resources import (
+from that_depends.entities.context import (
     ContextScope,
     ContextScopes,
-    DIContextMiddleware,
     InvalidContextError,
+    _enter_named_scope,
     get_current_scope,
 )
+from that_depends.entities.resource_context import ResourceContext
+from that_depends.meta import DefaultScopeNotDefinedError
+from that_depends.providers import DIContextMiddleware, container_context
 
 
 logger = logging.getLogger(__name__)
