+++ that_depends/container.py
@@ -121,21 +121,6 @@ class BaseContainer(metaclass=BaseContainerMeta):
             v.reset_override()
 
     @classmethod
-    def get_providers(cls) -> dict[str, AbstractProvider[typing.Any]]:
-        """Get all connected providers."""
-        if not hasattr(cls, "providers"):
-            cls.providers = {k: v for k, v in cls.__dict__.items() if isinstance(v, AbstractProvider)}
-        return cls.providers
-
-    @classmethod
-    def get_containers(cls) -> list[type["BaseContainer"]]:
-        """Get all connected containers."""
-        if not hasattr(cls, "containers"):
-            cls.containers = []
-
-        return cls.containers
-
-    @classmethod
     def resolver(cls, item: typing.Callable[P, T]) -> typing.Callable[[], typing.Awaitable[T]]:
         """Decorate a function to automatically resolve dependencies on call by name.
